+++ changelog.md
@@ -1,1 +1,2 @@
 * Updated to MC 1.21.5
+* Fixed problems with world type configuration screen